requirements.txt
--extra-index-url https://huggingface.github.io/autogptq-index/whl/cu118/

# START section of dependencies that don't install on Darwin/MacOS
bitsandbytes==0.46.1
# triton 3.4.0 is not compatible with CCE
triton>=3.0.0,<3.4.0
mamba-ssm==1.2.0.post1
xformers>=0.0.23.post1
autoawq==0.2.7.post3
liger-kernel==0.6.1
# END section

packaging==23.2

huggingface_hub>=0.33.0
peft==0.17.0
transformers==4.55.0
tokenizers>=0.21.1
accelerate==1.10.0
datasets==4.0.0
deepspeed>=0.17.0
trl==0.21.0
hf_xet==1.1.5
kernels==0.9.0
trackio

optimum==1.16.2
hf_transfer
sentencepiece
gradio==5.41.1

modal==1.0.2
pydantic==2.10.6
addict
fire
PyYAML>=6.0
requests
wandb
einops
colorama
numba
numpy>=1.24.4,<=2.0.1

# qlora things
evaluate==0.4.1
scipy
scikit-learn==1.4.2
nvidia-ml-py==12.560.30
art
tensorboard
python-dotenv==1.0.1

# remote filesystems
s3fs>=2024.5.0
gcsfs>=2024.5.0
adlfs>=2024.5.0
ocifs==1.3.2

zstandard==0.22.0
fastcore

# lm eval harness
lm_eval==0.4.7
langdetect==1.0.9
immutabledict==4.2.0
antlr4-python3-runtime==4.13.2

torchao==0.12.0
schedulefree==1.4.1

axolotl-contribs-lgpl==0.0.6
axolotl-contribs-mit==0.0.5

mistral-common==1.8.3
